+++ src/main/java/org/folio/dcb/integration/dcb/model/LocationAgenciesIds.java
@@ -1,3 +1,0 @@
-package org.folio.dcb.integration.dcb.model;
-
-public record LocationAgenciesIds(String institutionId, String campusId, String libraryId) {}