+++ src/main/java/org/folio/dcb/integration/dcb/model/LocationCodeNamePair.java
@@ -1,3 +1,0 @@
-package org.folio.dcb.integration.dcb.model;
-
-public record LocationCodeNamePair(String code, String name) {}